Move AddCollabrator functions to service layer
